some refactoring: - add placeholder fragments for Matches and Swiping tab - move fragments and adapter in packages
