+++ app/src/main/java/com/greenfox/opal/gitinder/MainActivity.java
@@ -11,9 +11,13 @@ import android.support.v4.view.ViewPager;
 import android.os.Bundle;
 import android.text.TextUtils;
 import android.util.Log;
+import com.greenfox.opal.gitinder.fragments.MatchesFragment;
+import com.greenfox.opal.gitinder.fragments.SettingsFragment;
+import com.greenfox.opal.gitinder.fragments.SwipingFragment;
 import com.greenfox.opal.gitinder.model.LoginRequest;
 import com.greenfox.opal.gitinder.model.response.LoginResponse;
 import com.greenfox.opal.gitinder.service.MockServer;
+import com.greenfox.opal.gitinder.service.SectionsPagerAdapter;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
