+++ app/src/main/java/com/greenfox/opal/gitinder/fragments/SwipingFragment.java
@@ -1,4 +1,4 @@
-package com.greenfox.opal.gitinder;
+package com.greenfox.opal.gitinder.fragments;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -6,13 +6,13 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import com.greenfox.opal.gitinder.R;
 
 public class SwipingFragment extends Fragment {
 
   @Nullable
   @Override
   public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-    View view = inflater.inflate(R.layout.settings_fragment, container, false);
-    return view;
+    return inflater.inflate(R.layout.fragment_swiping, container, false);
   }
 }
