+++ app/src/main/java/com/greenfox/opal/gitinder/service/SectionsPagerAdapter.java
@@ -1,4 +1,4 @@
-package com.greenfox.opal.gitinder;
+package com.greenfox.opal.gitinder.service;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
